updated routing, added to bottom section
