--- src/common/components/section/HomePage/HomeDoulaProgram.css
@@ -1,8 +1,0 @@
-.text-container {
-  display: flex;
-  flex-direction: column;
-  width: 90vw;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-}
