--- src/common/components/section/HomePage/HomeDoulaProgram.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import './HomeDoulaProgram.css';
-
-export default function HomeDoulaProgram() {
-  return (
-    <div>
-      <div className='text-container'>
-        <p>
-          SIU School of Medicine&apos;s Office of Certification Strategies
-          administers the Illinois Medicaid-Certified Doula Program with
-          oversight guidance by the Illinois Department of Healthcare and Family
-          Services.
-        </p>
-      </div>
-    </div>
-  );
-}
