+++ src/common/components/section/SectionComponent.css
@@ -4,6 +4,7 @@
   align-items: center;
   text-align: center;
   font-family: 'Poppins';
+  width: 100vw;
 }
 
 .page-section {
